requirements.txt
anyio==3.7.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
boto3==1.26.163
botocore==1.29.163
Brotli==1.0.9
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.98.0
fonttools==4.41.1
fpdf2==2.7.4
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.25.2
orjson==3.9.1
oscrypto==1.3.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
pdfkit==1.0.0
Pillow==10.0.0
platformdirs==3.8.0
playwright==1.36.0
pluggy==1.2.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.9
pydyf==0.7.0
pyee==9.0.4
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pyhtml2pdf==0.0.6
pypdf==3.14.0
PyPDF2==3.0.1
pyphen==0.14.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==7.4.0
pytest-base-url==2.0.0
pytest-playwright==0.3.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0
pyzipper==0.3.6
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
s3transfer==0.6.1
selenium==4.11.2
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starkbank-ecdsa==2.2.0
starlette==0.27.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.0
tzlocal==5.0.1
ujson==5.8.0
uritools==4.0.1
urllib3==1.26.16
uuid6==2023.5.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
weasyprint==59.0
webdriver-manager==4.0.0
webdrivermanager==0.10.0
webencodings==0.5.1
websockets==11.0.3
wsproto==1.2.0
xhtml2pdf==0.2.11
zopfli==0.2.2
